requirements.txt
asgiref==3.4.1
astroid==2.9.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
Deprecated==1.2.13
Django==3.2.5
django-ckeditor==6.2.0
django-debug-toolbar==3.2.3
django-js-asset==1.2.2
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mysqlclient==2.1.0
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.21
pylint==2.12.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
